+++ src/vs/platform/quickinput/browser/quickInput.ts
@@ -6,21 +6,18 @@
 import { IListRenderer, IListVirtualDelegate } from 'vs/base/browser/ui/list/list';
 import { List } from 'vs/base/browser/ui/list/listWidget';
 import { CancellationToken } from 'vs/base/common/cancellation';
-import { Codicon } from 'vs/base/common/codicons';
 import { IQuickInputOptions, IQuickInputStyles, QuickInputController } from 'vs/base/parts/quickinput/browser/quickInput';
-import { localize } from 'vs/nls';
 import { IAccessibilityService } from 'vs/platform/accessibility/common/accessibility';
 import { IContextKey, IContextKeyService, RawContextKey } from 'vs/platform/contextkey/common/contextkey';
-import { IInstantiationService, ServicesAccessor } from 'vs/platform/instantiation/common/instantiation';
+import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
 import { ILayoutService } from 'vs/platform/layout/browser/layoutService';
 import { IWorkbenchListOptions, WorkbenchList } from 'vs/platform/list/browser/listService';
 import { QuickAccessController } from 'vs/platform/quickinput/browser/quickAccess';
 import { IQuickAccessController } from 'vs/platform/quickinput/common/quickAccess';
-import { QuickPickItem, IInputBox, IInputOptions, IKeyMods, IPickOptions, IQuickInputButton, IQuickInputService, IQuickNavigateConfiguration, IQuickPick, IQuickPickItem, IQuickPickItemButtonEvent, QuickPickInput } from 'vs/platform/quickinput/common/quickInput';
-import { IStorageService, StorageScope, StorageTarget } from 'vs/platform/storage/common/storage';
+import { IInputBox, IInputOptions, IKeyMods, IPickOptions, IQuickInputButton, IQuickInputService, IQuickNavigateConfiguration, IQuickPick, IQuickPickItem, QuickPickInput } from 'vs/platform/quickinput/common/quickInput';
 import { activeContrastBorder, badgeBackground, badgeForeground, buttonBackground, buttonForeground, buttonHoverBackground, contrastBorder, inputBackground, inputBorder, inputForeground, inputValidationErrorBackground, inputValidationErrorBorder, inputValidationErrorForeground, inputValidationInfoBackground, inputValidationInfoBorder, inputValidationInfoForeground, inputValidationWarningBackground, inputValidationWarningBorder, inputValidationWarningForeground, keybindingLabelBackground, keybindingLabelBorder, keybindingLabelBottomBorder, keybindingLabelForeground, pickerGroupBorder, pickerGroupForeground, progressBarBackground, quickInputBackground, quickInputForeground, quickInputListFocusBackground, quickInputListFocusForeground, quickInputListFocusIconForeground, quickInputTitleBackground, widgetShadow } from 'vs/platform/theme/common/colorRegistry';
 import { computeStyles } from 'vs/platform/theme/common/styler';
-import { IThemeService, Themable, ThemeIcon } from 'vs/platform/theme/common/themeService';
+import { IThemeService, Themable } from 'vs/platform/theme/common/themeService';
 
 export interface IQuickInputControllerHost extends ILayoutService { }
 
@@ -229,73 +226,5 @@ export class QuickInputService extends Themable implements IQuickInputService {
 				pickerGroupForeground
 			})
 		};
-	}
-
-	formatPinnedItems(accessor: ServicesAccessor, storageKey: string, quickPick: IQuickPick<IQuickPickItem>, callback: () => void): void {
-		const storageService = accessor.get(IStorageService);
-		quickPick.onDidTriggerItemButton((e) => {
-			if (e.button.iconClass === ThemeIcon.asClassName(Codicon.pin) || e.button.iconClass === ThemeIcon.asClassName(Codicon.pinned)) {
-				this._formatPinnedItems(storageKey, quickPick, storageService, e, callback);
-				callback();
-			}
-		});
-		this._formatPinnedItems(storageKey, quickPick, storageService, undefined, callback);
-	}
-
-	private async _formatPinnedItems(storageKey: string, quickPick: IQuickPick<IQuickPickItem>, storageService: IStorageService, event?: IQuickPickItemButtonEvent<IQuickPickItem>, callback?: () => void): Promise<void> {
-		const formattedItems: QuickPickItem[] = [];
-		const labels = this._getPinnedLabels(storageKey, storageService);
-		const updatedLabels = !!event?.item.label ? this._updatePinned(storageKey, event.item.label, storageService, new Set(labels).has(event.item.label)) : labels.filter(l => l !== 'Pinned');
-		if (updatedLabels.length) {
-			formattedItems.push({ type: 'separator', label: localize("terminal.commands.pinned", 'Pinned') });
-		}
-		for (const label of updatedLabels) {
-			const item = quickPick.items.find(i => i.label === label && i.type !== 'separator');
-			if (item) {
-				const pinnedItem = Object.assign({}, item);
-				this._updateButtons(pinnedItem, false);
-				formattedItems.push(pinnedItem);
-			}
-		}
-
-		for (const item of quickPick.items.filter(i => !!i.label)) {
-			this._updateButtons(item, true);
-			formattedItems.push(item);
-		}
-		quickPick.items = formattedItems;
-		quickPick.onDidChangeValue(value => {
-			// don't show pinned items in the search results
-			quickPick.items = value ? quickPick.items.filter(i => i.type !== 'separator' && !i.buttons?.find(b => b.iconClass === ThemeIcon.asClassName(Codicon.pinned))) : quickPick.items;
-			if (!value && callback) {
-				callback();
-			}
-		});
-		quickPick.show();
-	}
-
-	private _updateButtons(item: QuickPickItem, removePin: boolean): void {
-		if (item.type === 'separator') {
-			return;
-		}
-		item.buttons = item.buttons ? item.buttons?.filter(b => b.iconClass !== ThemeIcon.asClassName(Codicon.pin) && b.iconClass !== ThemeIcon.asClassName(Codicon.pinned)) : [];
-		item.buttons.unshift({
-			iconClass: removePin ? ThemeIcon.asClassName(Codicon.pin) : ThemeIcon.asClassName(Codicon.pinned),
-			tooltip: removePin ? localize('pinCommand', "Pin command") : localize('pinnedCommand', "Pinned command"),
-			alwaysVisible: false
-		});
-	}
-
-	private _updatePinned(storageKey: string, label: string, storageService: IStorageService, removePin: boolean): string[] {
-		const pinnedLabels = this._getPinnedLabels(storageKey, storageService).filter(l => l !== label);
-		if (!removePin) {
-			pinnedLabels.push(label);
-		}
-		storageService.store(storageKey, JSON.stringify(pinnedLabels), StorageScope.WORKSPACE, StorageTarget.USER);
-		return pinnedLabels;
-	}
-
-	private _getPinnedLabels(storageKey: string, storageService: IStorageService): string[] {
-		const items = storageService.get(storageKey, StorageScope.WORKSPACE);
-		return items ? JSON.parse(items) : [];
 	}
 }
